clientside code refactor
